requirements.txt
ultralytics
streamlit
easyocr
opencv-python-headless
numpy
Pillow